third_party/blink/web_tests/editing/execCommand/indent-list-item-with-children-expected.txt


Indent With Children:
| "
    "
| <ol>
|   "
        "
|   <ol>
|     <li>
|       id="listItemBlock"
|       "before"
|       <div>
|         "div1"
|       <div>
|         "div2"
|       <div>
|         "div3"
|       "after"
|     <li>
|       id="listItemInline"
|       "before"
|       <span>
|         "span1"
|       <span>
|         "span2"
|       <span>
|         "span3"
|       "after"
|     <li>
|       id="listItemTable"
|       "before"
|       <table>
|         border="1"
|         cellpadding="2"
|         cellspacing="2"
|         <tbody>
|           <tr>
|             <td>
|               <br>
|       "after"
|   "
        "
|   "
        "
|   "
    "
| "
"

Indent With Pre:
| "
"
| <ol>
|   "
"
|   <pre>
|     <ul>
|       <ul>
|         <li>
|           id="listItemPre"
|           "<#selection-caret>hello"
|           "
"
|           "world"
|       "
"
|   "
"
| "
"
